implementation/src/main/java/com/antoniovinter/implementation/services/converter/UserConverter.java
package com.antoniovinter.implementation.services.converter;


import com.antoniovinter.api.dto.UserDto;
import com.antoniovinter.implementation.domain.User;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Component;

@Component
public class UserConverter {

    public UserDto UserToDto(User user){
        ModelMapper modelMapper = new ModelMapper();
        UserDto userDto = modelMapper.map(user, UserDto.class);
        return userDto;
    }

    public User DtoToUser(UserDto userDto){
        ModelMapper modelMapper = new ModelMapper();
        User user = modelMapper.map(userDto, User.class);
        return user;
    }
}
